blog/requirements.txt
Django==2.2
djoser==2.0.3
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
psycopg2==2.8.5
gunicorn==20.0.4
drf-tracking==1.5.0